chore: move constant data into seperate file
