+++ src/MyForm.component.jsx
@@ -11,39 +11,9 @@ import {
   DatePicker,
   TimePicker,
 } from "antd";
+import {COUNTRIES, COLORS, RACES, FOODS, INITIAL_VALUE} from './data/const';
 
 const { RangePicker } = DatePicker;
-
-const COUNTRIES = [
-  { name: "China", value: "china" },
-  { name: "U.S.A", value: "usa" },
-];
-const COLORS = [
-  { name: "Red", value: "red" },
-  { name: "Green", value: "green" },
-  { name: "Blue", value: "blue" },
-];
-const RACES = [
-  { name: "Asian", value: "asian" },
-  { name: "African", value: "african" },
-  { name: "Caucasian", value: "caucasian" },
-];
-const FOODS = [
-  { name: "Rice", value: "rice" },
-  { name: "Noodle", value: "noodle" },
-  { name: "Meat", value: "meat" },
-];
-const INITIAL_VALUE = {
-  name: "",
-  country: "",
-  colors: [],
-  race: "",
-  isSwitched: false,
-  foods: [],
-  date: null,
-  range: [],
-  time: null,
-};
 
 export default class MyForm extends Component {
   constructor(props) {
@@ -209,7 +179,7 @@ export default class MyForm extends Component {
                 offset: 6,
               }}
             >
-              <Button type="primary" htmlType="submit">
+              <Button type="primary" htmlType="submit" style={{marginRight:'10px'}}>
                 Submit
               </Button>
               <Button htmlType="button" onClick={this.handleReset}>Reset</Button>
